Change the section creation logic in add_url
Change the section creation logic in add_url. Each section
should be of the form YYYY-MM. When adding a url, the section
with the latest date is checked, if this is less thean the
current date then a new section with the current date is
created and the url is added to that. Otherwise the urls
are added to the section with the latest date.